+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add color picker in schedule chip to change course color
+- Add color picker (with its own component) in schedule chip to change course color
 
 ### Fixed